# greenlight-api/docs/openapi.json
{
    "components": {
        "schemas": {
            "Metadata": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "first_page": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "page_size": {
                        "type": "integer"
                    },
                    "total_records": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "Movie": {
                "properties": {
                    "genres": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "id": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "runtime": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "version": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "year": {
                        "format": "int32",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "RegisterUserPayload": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "activated": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "id": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "activateUserPayload": {
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "createMoviePayload": {
                "properties": {
                    "genres": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "runtime": {
                        "format": "int32",
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "year": {
                        "format": "int32",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "loginPayload": {
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "updateMoviePayload": {
                "properties": {
                    "genres": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    },
                    "runtime": {
                        "format": "int32",
                        "nullable": true,
                        "type": "integer"
                    },
                    "title": {
                        "nullable": true,
                        "type": "string"
                    },
                    "year": {
                        "format": "int32",
                        "nullable": true,
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        }
    },
    "info": {
        "contact": {
            "name": "Allen Francis"
        },
        "description": "\n\t\t\tGreenlight is an api for a service like IMDB, where users can\n\t\t\tadd, list and edit details about movies. I built this to learn building\n\t\t\tweb APIs in Go. The api OpenAPI API definition of this was created using \n\t\t\thttps://github.com/lallenfrancisl/gopi, a tool that I made. And the documentation\n\t\t\tUI is rendered using https://scalar.com\n\t\t",
        "license": {
            "name": ""
        },
        "title": "Greenlight movie database RESTful API",
        "version": "1.0.0"
    },
    "openapi": "3.0.0",
    "paths": {
        "/v1/healthcheck": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string"
                                        },
                                        "system_info": {
                                            "properties": {
                                                "environment": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "string"
                                                },
                                                "version": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "Healthcheck"
                ]
            }
        },
        "/v1/movies": {
            "get": {
                "operationId": "listAllTheMovies",
                "parameters": [
                    {
                        "description": "Search by title",
                        "in": "query",
                        "name": "title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by list of genres",
                        "in": "query",
                        "name": "genres",
                        "schema": {
                            "items": {
                                "type": "string"
                            },
                            "type": "array"
                        }
                    },
                    {
                        "description": "Page number",
                        "in": "query",
                        "name": "page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Number of items in each page",
                        "in": "query",
                        "name": "page_size",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Sort by given field name and direction",
                        "in": "query",
                        "name": "sort",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "metadata": {
                                            "$ref": "#/components/schemas/Metadata"
                                        },
                                        "movies": {
                                            "items": {
                                                "$ref": "#/components/schemas/Movie"
                                            },
                                            "type": "array"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "List all the movies",
                "tags": [
                    "Movies"
                ]
            },
            "post": {
                "operationId": "createANewMovie",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/createMoviePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movie": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Create a new movie",
                "tags": [
                    "Movies"
                ]
            }
        },
        "/v1/movies/{id}": {
            "delete": {
                "operationId": "deleteAMovieById",
                "parameters": [
                    {
                        "description": "Id of the movie",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Delete a movie by id",
                "tags": [
                    "Movies"
                ]
            },
            "get": {
                "operationId": "getAMovieById",
                "parameters": [
                    {
                        "description": "Id of the movie",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movie": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Get a movie by id",
                "tags": [
                    "Movies"
                ]
            },
            "patch": {
                "operationId": "updateAMovieById",
                "parameters": [
                    {
                        "description": "Id of the movie",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateMoviePayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movie": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Update a movie by id",
                "tags": [
                    "Movies"
                ]
            }
        },
        "/v1/users": {
            "post": {
                "description": "Create a new user and start the activation flow",
                "operationId": "createANewUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Create a new user",
                "tags": [
                    "Users"
                ]
            }
        },
        "/v1/users/login": {
            "post": {
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/loginPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "credentials": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Login",
                "tags": [
                    "Users"
                ]
            }
        },
        "/v1/users/{id}/activate": {
            "put": {
                "operationId": "activateUser",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/activateUserPayload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": ""
                    }
                },
                "summary": "Activate user",
                "tags": [
                    "Users"
                ]
            }
        }
    },
    "tags": [
        {
            "description": "Healthcheck routes",
            "externalDocs": {},
            "name": "Healthcheck"
        },
        {
            "description": "APIs for managing movies",
            "externalDocs": {},
            "name": "Movies"
        },
        {
            "description": "User managment APIs",
            "externalDocs": {},
            "name": "Users"
        }
    ]
}